requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
numpy
requests
Pillow
python-multipart